+++ icemet/src/misc.cpp
@@ -6,7 +6,7 @@ namespace cv { namespace icemet {
 
 void adjust(const cv::UMat& src, cv::UMat& dst, uchar a0, uchar a1, uchar b0, uchar b1)
 {
-	size_t gsize[1] = {src.cols * src.rows};
+	size_t gsize[1] = {(size_t)(src.cols * src.rows)};
 	ocl::Kernel("adjust", ocl::icemet::misc_oclsrc).args(
 		ocl::KernelArg::PtrReadOnly(src),
 		ocl::KernelArg::PtrWriteOnly(dst),
@@ -18,7 +18,7 @@ void hist(const cv::UMat& src, cv::Mat& counts, cv::Mat& bins, float min, float 
 {
 	int n = roundf((max-min) / step);
 	cv::UMat tmp = cv::UMat::zeros(1, n, CV_32SC1);
-	size_t gsize[1] = {src.cols * src.rows};
+	size_t gsize[1] = {(size_t)(src.cols * src.rows)};
 	ocl::Kernel("hist", ocl::icemet::misc_oclsrc).args(
 		ocl::KernelArg::PtrReadOnly(src),
 		ocl::KernelArg::PtrReadWrite(tmp),
@@ -34,7 +34,7 @@ void hist(const cv::UMat& src, cv::Mat& counts, cv::Mat& bins, float min, float 
 void imghist(const cv::UMat& src, cv::Mat& dst)
 {
 	cv::UMat tmp = cv::UMat::zeros(1, 256, CV_32SC1);
-	size_t gsize[1] = {src.cols * src.rows};
+	size_t gsize[1] = {(size_t)(src.cols * src.rows)};
 	ocl::Kernel("imghist", ocl::icemet::misc_oclsrc).args(
 		ocl::KernelArg::PtrReadOnly(src),
 		ocl::KernelArg::PtrReadWrite(tmp)
